add abc413

+++ abc413/abc413_c.cpp
@@ -7,16 +7,16 @@ int main() {
   cin >> Q;
   if ((Q < 1) || (Q > 2 * 10e5))
     exit(1);
-  vector<pair<long, long>> A;
-  int offset = 0;
+  vector<pair<int64_t, int64_t>> A;
+  int64_t offset = 0;
   rep (i, Q) {
     int j;
     cin >> j;
     if ((j != 1) && (j != 2))
       exit(1);
     if (j == 1) {
-      pair<long long, long long> p;
-      long long c, x;
+      pair<int64_t, int64_t> p;
+      int64_t c, x;
       cin >> c >> x;
       if ((c < 1) || (c > 10e9))
 	exit(1);
@@ -25,13 +25,13 @@ int main() {
       p = make_pair(c, x);
       A.push_back(p);
     } else { // j == 2
-      long long k;
+      int64_t k;
       cin >> k;
       if ((k < 1) || (k > 10e9))
 	exit(1);
-      int lower = offset;
+      int64_t lower = offset;
       offset = offset + k;
-      long long sum = 0;
+      int64_t sum = 0;
       int m = 0;
       while (m < A.size()) {
 	if (lower - A[m].first > 0)
@@ -40,14 +40,11 @@ int main() {
 	  break;
 	m++;
       }
-      //cout << "m = " << m << endl;
-      //cout << "lower = " << lower << endl;
       if (lower + k <= A[m].first) {
 	sum = A[m].second * k;
       } else {
 	sum = A[m].second * (A[m].first - lower);
 	k = k - (A[m].first - lower);
-	// cout << "sum =" << sum << endl;
 	m++;
 	while (k > 0) {
 	  if (k >= A[m].first) {
